refactor: remove context file
