+++ frontend/packages/app/src/pages/projects/list/index.tsx
@@ -1,10 +1,10 @@
 /**
  * Internal dependencies.
  */
-import ProjectsHeader from "../components/header";
-import { ProjectListSubHeader } from "../sub-header";
 import { ProjectListProvider } from "./provider";
 import ProjectListView from "./view";
+import ProjectsHeader from "../components/header";
+import { ProjectListSubHeader } from "../components/sub-header";
 
 function ProjectListPage() {
   return (
