+++ frontend/packages/app/src/pages/projects/utils.ts
@@ -1,0 +1,18 @@
+/**
+ * Internal dependencies.
+ */
+import type { ProjectListFilters } from "./types";
+
+export const buildListFrappeFilters = (filters: ProjectListFilters) => {
+  const out: unknown[] = [];
+  if (filters.ragStatus) {
+    out.push(["custom_project_rag_status", "=", filters.ragStatus]);
+  }
+  if (filters.phase) {
+    out.push(["custom_project_phase", "=", filters.phase]);
+  }
+  if (filters.status) {
+    out.push(["status", "=", filters.status]);
+  }
+  return out;
+};
